requirements.txt
torch
torchvision
torchaudio
scikit-learn
jupyterlab
jupytext
pandas
seaborn
umap-learn
colorcet
scikit-image
bokeh
holoviews
datashader
yellowbrick
lux-api